Glom onto some code for parsing arg spec strings.
This is just a simple lexer and tests for its tokenization,
and provides no value on its own.
It might be a useful component to further work.
There will be a lot to go in that, however;
among the things _not_ included here is any actual recognition
of recursive structures (nor even validating that tokens which
describe recursive structures actually match!);
and certainly not putting the results to work.

The original is hoisted from another open-source library,
as indicated in the comments in the content.

I've modified it a bit already in error handling,
and it the way it recognizes positional arguments
(these now require being bracketed by '<>' instead of
being required to be in capslock/screamcase).